requirement.txt
django-allauth==65.14.3
requests==2.32.3
pyjwt==2.11.0
cryptography==46.0.5
pillow==12.1.0
crispy-bootstrap4==2025.3
django-crispy-forms==2025.3